backend/requirement.txt
asgiref                 3.2.7
Django                  3.0.5
django-cors-headers     3.2.1
django-rest-auth        0.9.5
djangorestframework     3.11.0
djangorestframework-jwt 1.11.0
pip                     19.0.3
PyJWT                   1.7.1
pytz                    2019.3
setuptools              40.8.0
six                     1.14.0
sqlparse                0.3.1